use int32 as sample type int for libsnd

+++ src/InputLibraryWrapper/LibSNDWrapper.h
@@ -12,7 +12,7 @@
   *
   */
 
-class LibSNDWrapper : public StandardWrapper<float>
+class LibSNDWrapper : public StandardWrapper<int32_t>
 {
 public:
